+++ frontend/src/components/form/BirthField/BirthField.module.css
@@ -1,7 +1,6 @@
-.container {
+.box {
   display: flex;
   flex-direction: column;
-  margin: 0.625rem;
 }
 
 .birth {
