+++ frontend/src/pages/Login/Login.js
@@ -1,28 +1,28 @@
 import React from "react";
-import { Link, useHistory } from "react-router-dom";
-
+import { generatePath, Link, useHistory, useLocation } from "react-router-dom";
+import Button from "../../components/@common/Button/Button";
 import Container, {
   CONTAINER_SIZE,
 } from "../../components/@common/Container/Container";
-
+import Form from "../../components/form/Form/Form";
+import FormInput from "../../components/form/FormInput/FormInput";
+import SubmitButton from "../../components/form/SubmitButton";
+import { SUCCESS_MESSAGE } from "../../constants/messages";
+import PATH, { PARAM } from "../../constants/path";
 import useForm from "../../hooks/useForm";
 import useTokenContext from "../../hooks/useTokenContext";
 import FormProvider from "../../provider/FormProvider";
+import { generateQuery } from "../../utils/route/query";
 import { validateEmail } from "../../utils/validation/email";
 import { validatePassword } from "../../utils/validation/password";
-import { SUCCESS_MESSAGE } from "../../constants/messages";
-import PATH from "../../constants/path";
-
 import styles from "./Login.module.css";
-import Form from "../../components/form/Form/Form";
-import Button from "../../components/@common/Button/Button";
-import FormInput from "../../components/form/FormInput/FormInput";
-import SubmitButton from "../../components/form/SubmitButton";
 
 const Login = () => {
-  const { fetchLogin } = useTokenContext();
-
   const history = useHistory();
+  const location = useLocation();
+  const currentRecruitment = location.state?.currentRecruitment;
+
+  const { fetchLogin } = useTokenContext();
 
   const submit = async (value) => {
     try {
@@ -32,6 +32,20 @@ const Login = () => {
       });
 
       alert(SUCCESS_MESSAGE.API.LOGIN);
+
+      if (currentRecruitment) {
+        history.push({
+          pathname: generatePath(PATH.APPLICATION_FORM, {
+            status: PARAM.APPLICATION_FORM_STATUS.NEW,
+          }),
+          search: generateQuery({ recruitmentId: currentRecruitment.id }),
+          state: {
+            currentRecruitment,
+          },
+        });
+
+        return;
+      }
 
       history.push(PATH.RECRUITS);
     } catch (e) {
